slight corrections and a README.
a new README.  Not much to say here, but it tries to explain
what to do if you want to create a new photos html file, and to send
data to the proper place on the drive.

Also, modified the script so it doesn't delete the destination directory
first.  I am currently keeping photos in that directory so I don't
want to clobber them.